requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
streamlit==1.41.1
httpx==0.28.1
websockets==14.1
pandas==2.2.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-genai>=0.5.0
pydantic-settings==2.7.1
requests==2.32.3
eval-type-backport==0.2.2
gTTS==2.5.4
